backend/requirements.txt
asgiref==3.8.1
crispy-tailwind==1.0.3
Django==5.1.6
django-browser-reload==1.18.0
django-cors-headers==4.7.0
django-crispy-forms==2.3
django-debug-toolbar==6.0.0
django-tailwind==4.2.0
djangorestframework==3.15.2
sqlparse==0.5.3
tzdata==2025.1
